release 2 commits from zigwin32gen commit bc5add006555c4c332cce33e34b107ebd6e3f5b0
commit bc5add006555c4c332cce33e34b107ebd6e3f5b0
enums: remove `.fmt()` method in favor of the standard `.format()` method

Removes the `.fmt()` method which provides a custom formatter for enums
that leverages a switch to get the tag name in favor of providing a
`format` method on the type itself.

commit 023e3275918d30fb91a3f20d1f4279794a06f116
avoid storing over 100Kb of enum value tag names for WIN32_ERROR

Replaces the "fmt" function for the WIN32_ERROR enum with one that
calls FormatMessage to get the error string instead of printing
the tag name. WIN32_ERROR has over 3000 values and requires over 100Kb
to store all the tag names as strings at runtime.